fix 3 warnings
example\direct_tcpip.c(209): warning C4244: 'function': conversion from 'unsigned int' to 'u_short', possible loss of data
tests\session_fixture.c(96): warning C4013: 'getcwd' undefined; assuming extern returning int
tests\session_fixture.c(100): warning C4013: 'chdir' undefined; assuming extern returning int

+++ example/direct_tcpip.c
@@ -206,7 +206,7 @@ int main(int argc, char *argv[])
     }
 
     sin.sin_family = AF_INET;
-    sin.sin_port = htons(local_listenport);
+    sin.sin_port = htons((unsigned short)local_listenport);
     sin.sin_addr.s_addr = inet_addr(local_listenip);
     if(INADDR_NONE == sin.sin_addr.s_addr) {
         perror("inet_addr");
